+++ go.mod
@@ -9,6 +9,8 @@ require (
 	github.com/spf13/viper v1.6.2
 	github.com/tidwall/gjson v1.6.0
 	github.com/tidwall/pretty v1.0.1 // indirect
+	go.uber.org/zap v1.10.0
 	golang.org/x/net v0.0.0-20200324143707-d3edc9973b7e
+	gopkg.in/natefinch/lumberjack.v2 v2.0.0
 	gopkg.in/tucnak/telebot.v2 v2.0.0-20200328014118-dd123e949ee1
 )
